Combine web-forms package Vite/Vitest configs
This is consistent with all of the other Web Forms packages. This merges the Vitest config as-is, into the Vite config from #184 with slight adjustments:

- Give a name to express intent of branching on mode === ‘demo’
- Set up branchy config options early, assign to single return object of all combined config

I think this makes it more clear what is conditional in the config, and why it should be conditional.

Note: this _mostly satisfies_ our Vue/non-Vue build story. I would like to consider aligning the Vue/non-Vue approach with the Solid/non-Solid approach (see: packages/xforms-engine/vite.config.ts, and the env flag set in packages/xforms-engine/package.json scripts).
